[verified] Extract deployment core config module
