+++ deployment-core/go.mod
@@ -1,0 +1,5 @@
+module github.com/devopsellence/devopsellence/deployment-core
+
+go 1.26.0
+
+require gopkg.in/yaml.v3 v3.0.1
